Move all the dragonbox files into their own folder
