+++ include/boost/charconv/detail/dragonbox/dragonbox.hpp
@@ -23,7 +23,7 @@
 #define BOOST_CHARCONV_DETAIL_DRAGONBOX_HPP
 
 #include <boost/charconv/detail/config.hpp>
-#include <boost/charconv/detail/dragonbox_common.hpp>
+#include <boost/charconv/detail/dragonbox/dragonbox_common.hpp>
 #include <boost/charconv/detail/bit_layouts.hpp>
 #include <boost/charconv/detail/emulated128.hpp>
 #include <boost/charconv/chars_format.hpp>
